Get data

+++ re_forecast/data/manage_data_storage.py
@@ -13,20 +13,26 @@ def register_exists(register_path = DATA_ENERGY_PRODUCTION_REGISTER) -> bool:
     return os.path.isfile(register_path)
 
 
-def file_exists(ressource_nb: int,
-                start_date,
-                end_date,
-                eic_code,
-                production_type,
-                production_subtype,
-                creation_date: str,
-                metadata_fields = METADATA_ENERGY_PRODUCTION_FIELDS) -> bool:
-    """"""
+def gen_file_exists(ressource_nb: int,
+                    start_date: str | None,
+                    end_date: str | None,
+                    eic_code: str | None,
+                    production_type: str | None,
+                    production_subtype: str | None,
+                    metadata_fields = METADATA_ENERGY_PRODUCTION_FIELDS,
+                    return_csv_name = False
+                    ) -> bool:
+    """Return True if the file exists, False otherwise.
+    Any file not present in the register is considered as non existing."""
+
+    # Create the register if it doesn't exists
+    if not register_exists():
+        create_register()
 
     # Read the register
     register = read_register()
 
-    # Recreate the csv path and the hash_id
+    # Recreate the csv path
     csv_name = create_csv_path("", # It is just to trigger the create_csv_path function and avoid a positionnal argument error.
                                ressource_nb,
                                start_date,
@@ -35,16 +41,31 @@ def file_exists(ressource_nb: int,
                                production_type,
                                production_subtype,
                                return_csv_name = True)
-    hash_id = create_hash_id(creation_date, csv_name)
-    print(csv_name, hash_id)
 
-    # Check in the register if already exists, and if so return data
-    hash_col_name = metadata_fields[1]
+    # Check in the register if the file already exists, and if so return data
+    csv_name_key = metadata_fields[8]
 
     # Transform the query into a bool
-    query_response_state = register.query(f"{hash_col_name} == {hash_id}").empty
+    file_exists_bool = not register.query(f"{csv_name_key} == '{csv_name}'").empty
 
-    return True if not query_response_state else False
+    # If the return csv_name param is set to True, return file_exists and the csv name
+    if return_csv_name:
+        return file_exists_bool, csv_name
+
+    return file_exists_bool
+
+
+def units_names_file_exists(ressource_nb: int,
+                            root_data_path = DATA_CSV_ENERGY_PRODUCTION_PATH
+                            ) -> bool:
+    """Return true if the units names file exists, false
+    otherwise."""
+
+    # Retreive the csv path
+    units_names_path = create_csv_path_units_names(root_data_path,
+                                                   ressource_nb)
+
+    return os.path.isfile(units_names_path)
 
 
 def create_register(fields = METADATA_ENERGY_PRODUCTION_FIELDS,
@@ -71,7 +92,8 @@ def create_register(fields = METADATA_ENERGY_PRODUCTION_FIELDS,
 
 
 def create_hash_id(*params) -> int:
-    """Create a hash int given any number of params"""
+    """Create a hash int given any number of params.
+    This function is conserved for legacy but no longer used."""
 
     # Create an empty string as base to receive the params str
     hash_base = ""
@@ -108,10 +130,9 @@ def create_metadata_row(ressource_nb: int,
 
     # Pick the names of the fields to append to the 'metadata' dict
     # in the metadata fields general parameter
-    hash_key = metadata_fields[1]
-    creation_date_key = metadata_fields[2]
-    ressource_key = metadata_fields[3]
-    csv_name_key = metadata_fields[9]
+    creation_date_key = metadata_fields[1]
+    ressource_key = metadata_fields[2]
+    csv_name_key = metadata_fields[8]
 
     # Append to the row the creation date
     now_dt = datetime.datetime.now()
@@ -132,9 +153,6 @@ def create_metadata_row(ressource_nb: int,
                                return_csv_name = True)
     metadata[csv_name_key] = csv_name
 
-    # Create a hash with the csv name (and the curent date), and fill the hash_id field
-    metadata[hash_key] = create_hash_id(now_str, csv_name)
-
     return metadata
 
 
@@ -148,7 +166,7 @@ def fill_register(ressource_nb: int,
                   register_path = DATA_ENERGY_PRODUCTION_REGISTER
                   ) -> None:
     """Fill the register with data generation date, file name, ressource called
-    name, unique hash id and params values."""
+    name and params values."""
 
     # Use the create register function to create the register if it not already exists
     create_register()
@@ -178,37 +196,48 @@ def read_register(register_path = DATA_ENERGY_PRODUCTION_REGISTER) -> pd.DataFra
     return register
 
 
-def delete_generation_data(*hash_id: int,
+def delete_generation_data(ressource_nb: int,
+                           start_date: str | None,
+                           end_date: str | None,
+                           eic_code: str | None,
+                           production_type: str | None,
+                           production_subtype: str | None,
                            register_path = DATA_ENERGY_PRODUCTION_REGISTER,
                            root_data_path = DATA_CSV_ENERGY_PRODUCTION_PATH,
                            metadata_fields = METADATA_ENERGY_PRODUCTION_FIELDS
                            ) -> None:
-    """Remove the generation data csvs among the specified hash_id corresponding
-    to these csvs. Remove the lines corresponding to these files in the
-    register accordingly."""
+    """Remove a generation data file given its ressource number and its params,
+    and remove the corresponding row inside the register accordingly."""
 
-    # Open the register
-    register = pd.read_csv(register_path)
+    # Determine the existence of the file and recreate the csv file name
+    file_exists_bool, csv_name = gen_file_exists(ressource_nb,
+                                                 start_date,
+                                                 end_date,
+                                                 eic_code,
+                                                 production_type,
+                                                 production_subtype,
+                                                 return_csv_name = True)
 
-    # Iterate over the hash_id tuple
-    for hash in hash_id:
-        # Read the csv name
-        hash_col = metadata_fields[1]
-        csv_name_col = metadata_fields[9]
-        csv_name = register.loc[register[hash_col] == hash, csv_name_col]
+    # If the file you want to delete does not exists, print an error message
+    if not file_exists_bool:
+        print("The file you want to delete does not exists")
+        return
 
-        # Contruct the csv path and delete the file
+    else:
+        # Open the register
+        register = read_register()
+
+        # Remove the row corresponding to the csv_name
+        csv_name_col = metadata_fields[8]
+        row_index = register.loc[register[csv_name_col] == csv_name, :].index
+        new_register = register.drop(row_index, axis = 0)
+
+        # Construct the csv_path and delete the file
         csv_path = f"{root_data_path}/{csv_name}"
-
-        # Delete the csv file
         os.remove(csv_path)
 
-    # Delete the lines in the register df in one time
-    indexes = register.loc[register.loc[hash_col].isin(hash_id), :].index
-    new_register = register.drop(indexes)
-
-    # Save the new register, overwrite the old register csv
-    new_register.to_csv(register_path)
+        # Save the new register by overwritting
+        new_register.to_csv(register_path)
 
 
 def delete_units_names(*ressource_nb: int,
